se hace migracion a oracle 19c

+++ src/main/java/com/irojas/demojwt/User/UserRepository.java
@@ -4,6 +4,9 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
-    Optional<User> findByUsername(String username); 
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    // Spring Data genera:
+    // SELECT * FROM USERS u WHERE u.USERNAME = ? (respeta @Table y @Column)
+    Optional<User> findByUsername(String username);
 }
